bpo-40275: Move transient_internet from test.support to socket_helper (GH-19711)
